requirements.txt
Flask==2.3.4
python-docx==0.8.12
